Resolve null pointer when do Tables to delete

+++ src/main/resources/_core/database/mysql/stored-procedures/sp_xf_system_drop_all_tables_in_schema.sql
@@ -9,8 +9,11 @@ BEGIN
 
     DECLARE tables_count INT;
 
-    SELECT COUNT(1) INTO tables_count FROM information_schema.tables
-    WHERE TABLE_TYPE = 'BASE TABLE' AND TABLE_SCHEMA = database_name;
+    SELECT COUNT(1)
+    INTO tables_count
+    FROM information_schema.tables
+    WHERE TABLE_TYPE = 'BASE TABLE'
+      AND TABLE_SCHEMA = database_name;
 
     IF tables_count > 0 THEN
 
@@ -20,15 +23,19 @@ BEGIN
                      FROM information_schema.tables
                      WHERE TABLE_TYPE = 'BASE TABLE'
                        AND TABLE_SCHEMA = database_name
-                        AND TABLE_NAME REGEXP '^(mamba_|dim_|fact_|flat_)');
+                       AND TABLE_NAME REGEXP '^(mamba_|dim_|fact_|flat_)');
 
-        SET @drop_tables = CONCAT('DROP TABLE IF EXISTS ', @tbls);
+        IF (@tbls IS NOT NULL) THEN
 
-        SET foreign_key_checks = 0; -- Remove check, so we don't have to drop tables in the correct order, or care if they exist or not.
-        PREPARE drop_tbls FROM @drop_tables;
-        EXECUTE drop_tbls;
-        DEALLOCATE PREPARE drop_tbls;
-        SET foreign_key_checks = 1;
+            SET @drop_tables = CONCAT('DROP TABLE IF EXISTS ', @tbls);
+
+            SET foreign_key_checks = 0; -- Remove check, so we don't have to drop tables in the correct order, or care if they exist or not.
+            PREPARE drop_tbls FROM @drop_tables;
+            EXECUTE drop_tbls;
+            DEALLOCATE PREPARE drop_tbls;
+            SET foreign_key_checks = 1;
+
+        END IF;
 
     END IF;
